Adding `resolveConcreteIndexNames` to DataStreamsActionUtil, and moving it to server so it can be shared across plugins
